requirements.txt
Flask
pydantic
google-generativeai
